requirements.txt
# Зависимости для парсера Telegram
telethon==1.32.1  # Библиотека для работы с Telegram API
pandas==2.0.3     # Для работы с данными
python-dotenv==1.0.0  # Для загрузки переменных окружения
asyncio           # Для асинхронного программирования